Backend/data_engine/serializers.py
from rest_framework import serializers
from .models import Dataset


class DatasetSerializer(serializers.ModelSerializer):
    class Meta:
        model = Dataset
        fields = ['id', 'file', 'cleaned_file', 'uploaded_at', 'is_cleaned']
